Clamp, Weight, Effect - utility.clamp has been made more robust - currentweight now functions - Effect added, it's future is undecided - Metafile support added, even though there are none - dump.rdb for basic db data - updated readme - attributes now has a user

+++ Chaos/Objects/Monster.cs
@@ -21,7 +21,7 @@ namespace Chaos
         internal DateTime LastUpdate { get; set; }
         internal List<Item> Items { get; set; }
         internal uint Gold;
-        internal override byte HealthPercent => (byte)Utility.Clamp<uint>((CurrentHP * 100) / MaximumHP, 0, MaximumHP);
+        internal override byte HealthPercent => Utility.Clamp<byte>((int)((CurrentHP * 100) / MaximumHP), 0, (int)MaximumHP);
         internal override uint MaximumHP { get; }
         internal override uint CurrentHP { get; set; }
